requirements.txt
streamlit
streamlit-feedback
google-generativeai
python-dotenv
langchain
langsmith
langchain_google_genai
langchain_core
regex
IPython
pandas
google-play-scraper
beautifulsoup4
faiss-cpu

-e .